[ADD] -> Attester class created
