+++ example.env
@@ -1,2 +1,3 @@
 RPC=YOUR_RPC_ENDPOINT
 WALLET_PRIVATE_KEY=YOUR_PRIVATE_KEY
+TOKEN_BOT=YOUR_TELEGRAM_BOT_TOKEN
